@@ -1,4 +1,4 @@
-﻿using Amazon.CDK.AWS.SecretsManager;
+﻿ using Amazon.CDK.AWS.SecretsManager;
 using Amazon.CDK;
 using Constructs;
 using System;
@@ -7,37 +7,38 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Amazon.CDK.AWS.SSM;
+using System.Collections;
 
-namespace DemoInfrastructure.Constructs
+namespace DemoInfrastructure.Constructs;
+
+public class SecretsManagerConstruct : DemoConstructBase, IEnumerable<Secret>
 {
-    public class SecretsManagerConstruct : Construct
+
+    private List<Secret> Secrets {get;} = new ();
+    
+    public SecretsManagerConstruct(Construct scope, string id, SecretsManagerConstructProps props)
+        : base(scope, id)
     {
-        public SecretsManagerConstruct(Construct scope, string id, SecretsManagerConstructProps props)
-            : base(scope, id)
+        var secret1 = new Secret(this, "launchcode", new SecretProps
         {
-            var secret1 = new Secret(this, "launchcode", new SecretProps
-            {
-                SecretName = $"launchcode-secret-{props.EnvironmentPostFix}",
-                SecretStringValue = new SecretValue("12345"),
-                Description = "Launch Code"
-            });
-            CreateSecretOutput(secret1.Node.Id, secret1, "Launch Code");
-
-
-            var secret2 = new Secret(this, "secretplanlocation", new SecretProps
-            {
-                SecretName = $"secret-plan-location-secret-{props.EnvironmentPostFix}",
-                SecretStringValue = new SecretValue("Basement of the Alamo"),
-                Description = "Location of Secret Plan"
-            });
-            CreateSecretOutput(secret2.Node.Id, secret2, "Secret Plan Location");
-        }
-
-        private CfnOutput CreateSecretOutput(string id, Secret secret, string description = null) => new(this, $"{id}-secret", new CfnOutputProps
-        {
-            Value = secret.SecretName,
-            Description = $"{description} Secrets Manager Secret"
+            SecretName = $"launchcode-secret-{props.EnvironmentPostFix}",
+            SecretStringValue = new SecretValue("12345"),
+            Description = "Launch Code"
         });
+        CreateConstructOutput(secret1.Node.Id, () => secret1.SecretName, "Launch Code Secret");
+        Secrets.Add(secret1);
 
+        var secret2 = new Secret(this, "secretplanlocation", new SecretProps
+        {
+            SecretName = $"secret-plan-location-secret-{props.EnvironmentPostFix}",
+            SecretStringValue = new SecretValue("Basement of the Alamo"),
+            Description = "Location of Secret Plan"
+        });
+        CreateConstructOutput(secret2.Node.Id, () => secret2.SecretName, "Secret Plan Location Secret");
+        Secrets.Add(secret2);
     }
+
+    public IEnumerator<Secret> GetEnumerator() => Secrets.GetEnumerator();
+
+    IEnumerator IEnumerable.GetEnumerator() => Secrets.GetEnumerator();
 }
